+++ core/referral/helpers_for_test.go
@@ -1,3 +1,15 @@
+// Copyright (c) 2023 Gobalsky Labs Limited
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE.VEGA file and at https://www.mariadb.com/bsl11.
+//
+// Change Date: 18 months from the later of the date of the first publicly
+// available Distribution of this version of the repository, and 25 June 2022.
+//
+// On the date above, in accordance with the Business Source License, use
+// of this software will be governed by version 3 or later of the GNU General
+// Public License.
+
 package referral_test
 
 import (
@@ -5,16 +17,41 @@ import (
 	"testing"
 	"time"
 
+	"code.vegaprotocol.io/vega/core/events"
+	"code.vegaprotocol.io/vega/core/integration/stubs"
 	"code.vegaprotocol.io/vega/core/referral"
 	"code.vegaprotocol.io/vega/core/referral/mocks"
+	"code.vegaprotocol.io/vega/core/snapshot"
+	"code.vegaprotocol.io/vega/core/stats"
 	"code.vegaprotocol.io/vega/core/types"
+	"code.vegaprotocol.io/vega/logging"
+	"code.vegaprotocol.io/vega/paths"
 	typespb "code.vegaprotocol.io/vega/protos/vega"
 	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 type testEngine struct {
-	engine *referral.Engine
+	engine *referral.SnapshottedEngine
 	broker *mocks.MockBroker
+}
+
+func newSnapshotEngine(t *testing.T, vegaPath paths.Paths, now time.Time, engine *referral.SnapshottedEngine) *snapshot.Engine {
+	t.Helper()
+
+	log := logging.NewTestLogger()
+	timeService := stubs.NewTimeStub()
+	timeService.SetTime(now)
+	statsData := stats.New(log, stats.NewDefaultConfig())
+	config := snapshot.DefaultConfig()
+
+	snapshotEngine, err := snapshot.NewEngine(vegaPath, config, log, timeService, statsData.Blockchain)
+	require.NoError(t, err)
+
+	snapshotEngine.AddProviders(engine)
+
+	return snapshotEngine
 }
 
 func newEngine(t *testing.T) *testEngine {
@@ -27,7 +64,7 @@ func newEngine(t *testing.T) *testEngine {
 
 	broker := mocks.NewMockBroker(ctrl)
 
-	engine := referral.NewEngine(epochEngine, broker)
+	engine := referral.NewSnapshottedEngine(epochEngine, broker)
 
 	return &testEngine{
 		engine: engine,
@@ -43,6 +80,34 @@ func endEpoch(t *testing.T, ctx context.Context, te *testEngine, endTime time.Ti
 		EndTime: endTime,
 	})
 	te.engine.OnEpoch(ctx, types.Epoch{
-		Action: typespb.EpochAction_EPOCH_ACTION_START,
+		StartTime: endTime.Add(1 * time.Second),
+		Action:    typespb.EpochAction_EPOCH_ACTION_START,
 	})
+}
+
+func expectReferralProgramStartedEvent(t *testing.T, engine *testEngine) {
+	t.Helper()
+
+	engine.broker.EXPECT().Send(gomock.Any()).Do(func(evt events.Event) {
+		_, ok := evt.(*events.ReferralProgramStarted)
+		assert.True(t, ok, "Event should be a ReferralProgramStarted, but is %T", evt)
+	}).Times(1)
+}
+
+func expectReferralProgramEndedEvent(t *testing.T, engine *testEngine) *gomock.Call {
+	t.Helper()
+
+	return engine.broker.EXPECT().Send(gomock.Any()).Do(func(evt events.Event) {
+		_, ok := evt.(*events.ReferralProgramEnded)
+		assert.True(t, ok, "Event should be a ReferralProgramEnded, but is %T", evt)
+	}).Times(1)
+}
+
+func expectReferralProgramUpdatedEvent(t *testing.T, engine *testEngine) *gomock.Call {
+	t.Helper()
+
+	return engine.broker.EXPECT().Send(gomock.Any()).Do(func(evt events.Event) {
+		_, ok := evt.(*events.ReferralProgramUpdated)
+		assert.True(t, ok, "Event should be a ReferralProgramUpdated, but is %T", evt)
+	}).Times(1)
 }