+++ core/types/snapshot_nodes.go
@@ -909,6 +909,12 @@ func PayloadFromProto(p *snapshot.Payload) *Payload {
 		ret.Data = PayloadTeamSwitchesFromProto(dt)
 	case *snapshot.Payload_Vesting:
 		ret.Data = PayloadVestingFromProto(dt)
+	case *snapshot.Payload_CurrentReferralProgram:
+		ret.Data = PayloadCurrentReferralProgramFromProto(dt)
+	case *snapshot.Payload_NewReferralProgram:
+		ret.Data = PayloadNewReferralProgramFromProto(dt)
+	default:
+		panic(fmt.Errorf("missing support for payload %T", dt))
 	}
 
 	return ret
@@ -1074,6 +1080,12 @@ func (p Payload) IntoProto() *snapshot.Payload {
 		ret.Data = dt
 	case *snapshot.Payload_Vesting:
 		ret.Data = dt
+	case *snapshot.Payload_CurrentReferralProgram:
+		ret.Data = dt
+	case *snapshot.Payload_NewReferralProgram:
+		ret.Data = dt
+	default:
+		panic(fmt.Errorf("missing support for payload %T", dt))
 	}
 	return &ret
 }